+++ app/Livewire/Admin/Prestasi/EditPrestasi.php
@@ -25,6 +25,7 @@ class EditPrestasi extends Component
     public $foto;
     public $currentFoto;
     public $prestasiId;
+    public $oldData;
 
     public function mount($id)
     {
@@ -33,34 +34,53 @@ class EditPrestasi extends Component
         $this->description = $prestasi->description;
         $this->currentFoto = $prestasi->foto;
         $this->prestasiId = $prestasi->id;
+        $this->oldData = [
+            'title' => $prestasi->title,
+            'description' => $prestasi->description,
+            'currentFoto' => $prestasi->currentFoto,
+        ];
+        // dd($this->oldData);
     }
 
     public function update()
     {
-        $validate = $this->validate();
         // dd($this->all());
-        $prestasi = Prestasi::find($this->prestasiId);
+        $this->validate();
+        $newData = [
+            'title' => $this->title,
+            'description' => $this->description,
+            'currentFoto' => $this->foto,
+        ];
 
-        if ($this->foto) {
-            // Hapus foto lama jika ada
-            if ($this->currentFoto && File::exists(public_path('uploads/prestasi/' . $this->currentFoto))) {
-                File::delete(public_path('uploads/prestasi/' . $this->currentFoto));
+        if ($newData !== $this->oldData) {
+            // dd($newData);
+            $prestasi = Prestasi::find($this->prestasiId);
+            if ($this->foto) {
+                // Hapus foto lama jika ada
+                if ($this->currentFoto && File::exists(public_path('uploads/prestasi/' . $this->currentFoto))) {
+                    File::delete(public_path('uploads/prestasi/' . $this->currentFoto));
+                }
+
+                $filename = time() . '.' . $this->foto->getClientOriginalExtension();
+                $this->foto->storeAs('/prestasi', $filename, 'public_uploads');
+                $prestasi->foto = $filename;
+                $this->currentFoto = $filename;
             }
 
-            $filename = time() . '.' . $this->foto->getClientOriginalExtension();
-            $this->foto->storeAs('/prestasi', $filename, 'public_uploads');
-            $prestasi->foto = $filename;
-            $this->currentFoto = $filename;
+            $prestasi->title = $this->title;
+            $prestasi->description = $this->description;
+            $prestasi->save();
+            notyf()->position('y', 'top')->duration(3000)->ripple(true)->dismissible(true)->addSuccess('Data successfully updated');
+            return $this->redirect('/panel-admin/prestasi', navigate: true);
+        } else {
+            // dd(123);
+            notyf()->position('y', 'top')->duration(3000)->ripple(true)->dismissible(true)->addInfo('No data updates');
+            return $this->redirect('/panel-admin/prestasi', navigate: true);
         }
-
-        $prestasi->title = $this->title;
-        $prestasi->description = $this->description;
-        $prestasi->save();
-        return $this->redirect('/panel-admin/prestasi', navigate: true);
     }
 
     public function render()
     {
-        return view('livewire.admin.prestasi.edit-prestasi',);
+        return view('livewire.admin.prestasi.edit-prestasi', );
     }
 }
